fix: fix controllers

+++ src/core/groups/controllers/groups.controller.ts
@@ -1,18 +1,20 @@
 import {
-  Controller,
-  Post,
+  BadRequestException,
   Body,
+  Controller,
   Delete,
+  Get,
   HttpCode,
   Param,
-  Get,
+  Post,
 } from '@nestjs/common';
+import { isValidObjectId } from 'mongoose';
+import { CreateJoinCodeDto } from '../../join-codes/dtos/create-join-code.dto';
+import { JoinCodeDto } from '../../join-codes/dtos/join-code.dto';
 import { CodesService } from '../../join-codes/service/code.service';
-import { GroupsService } from '../service/groups.service';
 import { CreateGroupDto } from '../dtos/create-group.dto';
 import { GroupDto } from '../dtos/group.dto';
-import { JoinCodeDto } from '../../join-codes/dtos/join-code.dto';
-import { CreateJoinCodeDto } from '../../join-codes/dtos/create-join-code.dto';
+import { GroupsService } from '../service/groups.service';
 
 @Controller('groups')
 export class GroupsController {
@@ -21,26 +23,34 @@ export class GroupsController {
     private readonly codeService: CodesService,
   ) {}
 
-  @Get(':id')
-  async getGroup(@Param('id') id: string): Promise<GroupDto> {
-    return this.groupsService.get(id);
-  }
-
   @Get()
   async find(): Promise<GroupDto[]> {
     return this.groupsService.find();
   }
+
+  @Get(':id')
+  async get(@Param('id') id: string): Promise<GroupDto> {
+    if (!isValidObjectId(id)) {
+      throw new BadRequestException(
+        `Invalid group ID, should be valid ObjectId`,
+      );
+    }
+
+    return this.groupsService.get(id);
+  }
+
+  @Post()
+  @HttpCode(201)
+  create(@Body() createGroupDto: CreateGroupDto): Promise<GroupDto> {
+    return this.groupsService.create(createGroupDto);
+  }
+
   @Post('code')
   @HttpCode(201)
   createJoinCode(
     @Body() createJoinCodeDto: CreateJoinCodeDto,
   ): Promise<JoinCodeDto> {
     return this.codeService.create(createJoinCodeDto);
-  }
-  @Post()
-  @HttpCode(201)
-  create(@Body() createGroupDto: CreateGroupDto): Promise<GroupDto> {
-    return this.groupsService.createGroup(createGroupDto);
   }
 
   @Post('join/:code')
@@ -53,11 +63,24 @@ export class GroupsController {
     @Body() body: { userId: string; groupId: string },
   ): Promise<void> {
     const { userId, groupId } = body;
+    if (!isValidObjectId(userId) || !isValidObjectId(groupId)) {
+      throw new BadRequestException(
+        `Invalid groupId or userId, should be valid ObjectId`,
+      );
+    }
+
     return this.codeService.removeStudentFromGroup(groupId, userId);
   }
 
   @Delete(':idOrCode')
+  @HttpCode(204)
   remove(@Param('idOrCode') id: string): Promise<void> {
+    if (!isValidObjectId(id)) {
+      throw new BadRequestException(
+        `Invalid group ID, should be valid ObjectId`,
+      );
+    }
+
     return this.groupsService.remove(id);
   }
 }
